added comment array in articles, added newComment route
